pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.107.3",
    "pandas>=2.3.2",
    "pdfplumber>=0.11.7",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "pyyaml>=6.0.2",
    "reportlab>=4.4.3",
    "weasyprint>=66.0",
]
